Add 2 new methods of fetching RSS feeds
Since both YQL and cors.io stopped working, we lost the ability to fetch RSS
feeds on the proxy-less hosting of Delayed Replay on GitHub Pages. As discussed
in the research on issue #4, we found two additional alternatives:

* https://api.rss2json.com
* https://cloudquery.t9t.io

This change adds support for both of them, with a fallback from one on another
one if the first one fails. This change also implies that fetch methods now
return a boolean that signifies their success status.

Tested manually by loading the file directly in the browser as well as with a
local server (which emulates how GitHub Pages will host it) to ensure it works
in both modes.

Fixes https://github.com/mbrukman/delayed-replay/issues/4