Enable users to watch vodcasts via <video> tag
More podcasts are serving video files, not just audio files, and we have been
seeing missing controls in those cases, though the media plays audio. Turns out,
for showing playback controls, we have been checking for presence of audio
files, but then assigning a video URL to an audio control, so it was playing the
audio portion of the file, but we have not been seeing the video.

This fixes the issue by treating audio and video transparently as "media" files,
and handling video files specially by creating a <video> tag when it's present
to show the video stream.

Also added cleanup code to reset the state between podcasts, so that if the
previous episode was a video one and the current one is audio-only, we don't
show a non-functional video playback control which will disappear upon clicking
on an audio-only episode next.

Also adds an exception URL sanitization via `$sce.trustAsResourceUrl()`:
https://code.angularjs.org/1.5.8/docs/api/ng/service/$sce#trustAsResourceUrl

Fixes https://github.com/mbrukman/delayed-replay/issues/5